+++ front/src/components/user_card/user_card.vue
@@ -156,7 +156,7 @@ const UserCard = {
     },
     userProfileLink (user) {
       return generateProfileLink(
-        user.id, user.screen_name,
+        user.flakeId, user.screen_name,
         this.$store.state.instance.restrictedNicknames
       )
     }